# openapi/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.3
  title: Home Project Blog Service
  description: |
    Blog for the Home Project engineers
  license:
    name: MIT
    url: https://github.com/Home-Project-Engineering/home-project-blog/blob/main/LICENSE
servers:
  - url: http://localhost:8080/api/1
security:
  - basicAuth: []
...
paths:
...
  /posts:
    post:
      summary: Create Post
      description: Create a new Post.
      tags:
        - Posts
      operationId: createPost
      requestBody:
        $ref: '#/components/requestBodies/PostBody'
      responses:
        "201":
          $ref: '#/components/responses/PostResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "default":
          $ref: '#/components/responses/UnknownError'
    get:
      summary: Get posts
      description: Uses query keys to retrieve a subset of existing posts.
      tags:
        - Posts
      operationId: getPosts
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/tag_id'
        - $ref: '#/components/parameters/tag_name'
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/posts_sort'
        - $ref: '#/components/parameters/page_num'
        - $ref: '#/components/parameters/page_size'
      responses:
        "200":
          $ref: '#/components/responses/PostsResponse'
        "default":
          $ref: '#/components/responses/UnknownError'
  /posts/{id}:
    parameters:
      - $ref: '#/components/parameters/p_id'
    get:
      tags:
        - Posts
      summary: Get Post by ID
      description: Retrieves an existing Post associated with the specified ID.
      operationId: getPost
      responses:
        "200":
          $ref: '#/components/responses/PostResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "default":
          $ref: '#/components/responses/UnknownError'
    put:
      tags:
        - Posts
      summary: Update Post
      description: Updates information on a Post that exists in the system by the specified ID.
      operationId: updatePost
      requestBody:
        $ref: '#/components/requestBodies/PostBody'
      responses:
        "200":
          $ref: '#/components/responses/PostResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "default":
          $ref: '#/components/responses/UnknownError'
    delete:
      tags:
        - Posts
      summary: Remove Post
      description: Deletes a Post associated with a specified ID.
      operationId: removePost
      responses:
        "204":
          $ref: '#/components/responses/NoContentResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "default":
          $ref: '#/components/responses/UnknownError'
  /posts/{post_id}/comments:
    parameters:
      - $ref: '#/components/parameters/p_post_id'
    get:
      summary: Get comments
      description: Uses query keys to retrieve a subset of existing comments.
      tags:
        - Comments
      operationId: getComments
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/user_name'
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/comments_sort'
        - $ref: '#/components/parameters/page_num'
        - $ref: '#/components/parameters/page_size'
      responses:
        "200":
          $ref: '#/components/responses/CommentsResponse'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "default":
          $ref: '#/components/responses/UnknownError'
    post:
      summary: Create Comment
      description: Create a new Comment.
      tags:
        - Comments
      operationId: createComment
      requestBody:
        $ref: '#/components/requestBodies/CommentBody'
      responses:
        "201":
          $ref: '#/components/responses/CommentResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "default":
          $ref: '#/components/responses/UnknownError'
  /posts/{post_id}/comments/{id}:
    parameters:
      - $ref: '#/components/parameters/p_post_id'
      - $ref: '#/components/parameters/p_id'
    get:
      tags:
        - Comments
      summary: Get Comment by ID.
      description: Retrieves an existing Comment associated with the specified ID.
      operationId: getComment
      responses:
        "200":
          $ref: '#/components/responses/CommentResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "default":
          $ref: '#/components/responses/UnknownError'
    put:
      tags:
        - Comments
      summary: Update Comment
      description: Updates information on a Comment that exists in the system by the specified ID and post ID.
      operationId: updateComment
      requestBody:
        $ref: '#/components/requestBodies/CommentBody'
      responses:
        "200":
          $ref: '#/components/responses/CommentResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "default":
          $ref: '#/components/responses/UnknownError'
    delete:
      tags:
        - Comments
      summary: Remove Comment
      description: Deletes a Comment associated with a specified ID.
      operationId: removeComment
      responses:
        "204":
          $ref: '#/components/responses/NoContentResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "default":
          $ref: '#/components/responses/UnknownError'
...
components:
  headers:
    X-Total-Count:
      description: The total results for the particular query.
      schema:
        type: number
        format: int64
  schemas:
...
    Post:
      description: A representation of Post entity.
      type: object
      required:
        - title
        - body
        - previewAttachment
      properties:
        id:
          description: The ID of the Post.
          type: number
          format: int64
          readOnly: true
        tags:
          description: The subset of assosiated tags to the Post.
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        createdOn:
          description: The Post creation date.
          type: string
          example: "2017-07-21T17:32:28Z"
          format: date-time
          readOnly: true
        user:
          description: The creator of the Post.
          readOnly: true
          allOf:
            - $ref: "#/components/schemas/User"
        text:
          description: The content of the Post.
          type: string
        title:
          description: The summary of the Post.
          type: string
        previewAttachment:
          description: URL of the resource that will be used as an preview of the Post.
          type: string
        updatedOn:
          description: The Post update date.
          type: string
          example: "2017-07-21T17:32:28Z"
          format: date-time
          readOnly: true
...
    Tag:
      required:
        - name
      type: object
      properties:
        id:
          description: The ID of the Tag.
          type: number
          format: int64
          readOnly: true
        name:
          description: The name of the Tag.
          type: string
          example: Java8
...
    Error:
      type: object
      description: Representation of an Error that can appear using the application.
      properties:
        code:
          description: The code of an error that describes the Error.
          type: string
          readOnly: true
        message:
          description: The message of an error that describes the Error.
          type: string
          readOnly: true
  responses:
...
    NoContentResponse:
      description: The request was succesfully processed.
...
    CommentResponse:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Comment"
    PostResponse:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Post"
...
    CommentsResponse:
      description: OK
      headers:
        X-Total-Count:
          $ref: '#/components/headers/X-Total-Count'
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Comment"
...
    PostsResponse:
      description: OK
      headers:
        X-Total-Count:
          $ref: '#/components/headers/X-Total-Count'
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Post"
...
    ValidationError:
      description: The payload contains an error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    UnknownError:
      description: The unknown error appeard. Check your payload or contact support.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  requestBodies:
    CommentBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Comment"
    PostBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Post"
...
  parameters:
    p_id:
      in: path
      name: id
      schema:
        type: number
        format: int64
      required: true
    p_post_id:
      in: path
      name: post_id
      schema:
        type: number
        format: int64
      required: true
    page_num:
      in: query
      name: page_num
      schema:
        type: integer
    page_size:
      in: query
      name: page_size
      schema:
        type: integer
    id:
      in: query
      name: id
      schema:
        type: number
        format: int64
...
    user_name:
      in: query
      name: user_name
      schema:
        type: string
    tag_name:
      in: query
      name: tag_name
      schema:
        type: string
    user_id:
      in: query
      name: user_id
      schema:
        type: string
    tag_id:
      in: query
      name: tag_id
      schema:
        type: string
...
    posts_sort:
      in: query
      name: sort
      description: |
        In order to execute *asc*, you need to specify in the search *id* or *title* parameter.
        In order to complete the *desc* sorting must be specified in the query parameter *-id*
      schema:
        type: string
        enum: [ "id", "-id", "title", "-title" ]
        default: "-id"
    comments_sort:
      in: query
      name: sort
      description: |
        In order to execute *asc*, you need to specify in the search *id* parameter.
        In order to complete the *desc* sorting must be specified in the query parameter *-id*
      schema:
        type: string
        enum: [ "id", "-id" ]
        default: "-id"
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic